backend/requirements.txt
# python 3.11
# 在/backend/目录下运行为 pip install -r requirements.txt
aiofiles==24.1.0
annotated-types==0.7.0
anyio==4.8.0
assemblyai==0.36.0
bcrypt==4.2.1
black==25.1.0
blinker==1.9.0
Bottleneck==1.4.2
cachetools==5.5.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
contourpy==1.3.1
cryptography==44.0.1
cycler==0.12.1
decorator==5.1.1
dnspython==2.7.0
ecdsa==0.19.0
email_validator==2.2.0
et_xmlfile==2.0.0
fastapi==0.115.8
filetype==1.2.0
flake8==7.1.2
Flask==3.1.0
fonttools==4.55.6
google-api-core==2.24.0
google-auth==2.37.0
google-cloud-speech==2.30.0
googleapis-common-protos==1.66.0
grpcio==1.69.0
grpcio-status==1.69.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
imageio==2.36.1
imageio-ffmpeg==0.6.0
iniconfig==2.0.0
intel-cmplr-lib-ur
intel-openmp
isort==6.0.0
itsdangerous==2.2.0
Jinja2==3.1.5
joblib==1.4.2
kiwisolver==1.4.8
Markdown==3.7
MarkupSafe==3.0.2
matplotlib==3.10.0
mccabe==0.7.0
mkl==2025.0.1
mkl-service==2.4.1
mkl_fft==1.3.11
mkl_random==1.2.8
motor==3.7.0
moviepy==2.1.2
mypy-extensions==1.0.0
numexpr==2.10.1
numpy==1.26.4
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
passlib==1.7.4
pathspec==0.12.1
patsy==1.0.1
pillow==10.4.0
platformdirs==4.3.6
pluggy==1.5.0
proglog==0.1.10
proto-plus==1.25.0
protobuf==5.29.3
pyasn1
pyasn1_modules
pycodestyle==2.12.1
pycparser==2.22
pydantic==2.10.5
pydantic-settings==2.7.1
pydantic_core==2.27.2
pydub==0.25.1
pyflakes==3.2.0
PyJWT==2.8.0
pymongo==4.11.2
pyparsing==3.2.1
PySocks==1.7.1
pytest==8.3.4
pytest-asyncio==0.25.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.4.0
python-magic==0.4.27
python-multipart==0.0.20
pytz==2024.1
PyYAML==6.0.2
requests==2.32.3
rsa==4.9
scikit-learn==1.6.1
scipy==1.15.1
six==1.16.0
sniffio==1.3.1
socks==0
SpeechRecognition==3.14.1
starlette==0.45.3
statsmodels==0.14.4
tbb==2022.0.0
tcmlib==1.2.0
threadpoolctl==3.5.0
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2023.3
umf==0.9.1
urllib3==2.3.0
uvicorn==0.34.0
websockets==14.1
Werkzeug==3.1.3
zhipuai==2.1.5.20250106
pydantic-extra-types==2.5.0

# 日志系统相关依赖 - 不使用OpenTelemetry，避免与protobuf版本冲突
structlog==24.1.0  # 结构化日志库，用于生成结构化日志输出
python-json-logger==2.0.7  # JSON格式日志处理
loguru==0.7.2  # 简单而强大的日志库，替代OpenTelemetry
rich==13.7.1  # 增强终端输出，用于开发环境中的日志美化

# Redis相关依赖
redis==5.0.1
aioredis==2.0.1
 